Abstracted RESTful tasks to a new RESTfulTestCase.
